Improve/fix environment variables for Pod/services
This will differentiate the address on which Pod listens to requests,
and the Pod address that services will see in order to make the callback.

This commit only formalizes the difference and makes it clearer,
it was partly already the case before.

Additionally, default port value of 3030 was moved to a constant,
and docs were improved.